Logrotate for verify-fuel-docs
For every doc build we store all build results, html and pdf,
which take about ~65M per build.

We wanted to use artifactNumToKeep parameter, so we could keep
the logs for longer time but remove unneeded build result.

Unfortunately, it appears that HTML data attached to the build
is considered to be 'report' and not 'artifact'. So it can only
be rotated together with the build logs.

Therefore, we need to use the numToKeep variable, which will
limit the overall number of stored builds.

Change-Id: I0fa0ed682b9c1bbefa53a3db61c68fe70ac3fad9