Add customizable mirror to custom bvt
Add parameters: UBUNTU_MIRROR_ID and MIRROR_UBUNTU

Second parameter has a priority, and if set will be used as it is.
If MIRROR_UBUNTU is not set, we can use generic 'latest' or
'latest-stable' values for UBUNTU_MIRROR_ID or specify the exact
'ubuntu-<timestamp>' value.

Also add EXTRA_DEB_REPOS and EXTRA_DEB_REPOS_PRIORITY for custom
packages.

Change-Id: Ifbd7c767bc68491656002ac2e46773b4c02583fd