Search packages by "Provides" metadata element
As some packages in requirements-rpm.txt could be listed as
capabilities, not as exact package names, let's search for
"Provides" metadata element. I.e. the "fuel-library6.1"
package provides the "fuel-library" capability, so to avoid
hardcoding the version it would be better to list it in
requirements-rpm.txt as "fuel-library".

Also, some syntax polishing to comply to the shellcheck.

Change-Id: I7b2bbf3c83a91faef616f712b143a6e8133d6004