Updated job-template for custom scripts infrastructure
Added separated project to configure the job
for 'lp-assigner' script.

Updated job-template for all jobs - added 'inject' option to
'builders' section.

It needs to have an ability more effective configure 'lp-assigner' job,
for example, because there are a couple of string variables used in the
script. Those variables shouldn't be used in others scripts.

Change-Id: I307edb3d07c96f7ff546901135f037152cc7dd9b
Dependency-Id: Icf71c722c442923c9144d03a6aff3623141db571