Add update-jenkins-jobs job
Job is triggered on merge event to jenkins-jobs repo only for
changes to corresponding servers/${CI_NAME} directory
(or changes to 'common' dir).
Credentials are stored as global passwords in Jenkins masters,
jenkins_jobs.ini is placed outside of job workspace and is cleaned
up after build run (even on build abort).

This commit enables the job for infra-ci

Change-Id: I6f79587ec2d0e40f82ebfbb28d5cd059743fbe5c