Add quiet period for publish jobs
Publish jobs are usually run very fast and consequently
on one Jenkins slave. And on this particular slave we get the
"Too many open files" error.

We suppose that Jenkins java process might need some additional
time for garbage collection when it finishes the job, so we
propose to add some quiet period (1 minute for now) and monitor
slave behaviour for some time.

Change-Id: I63fcb6ff5cf81c16d2a468499ba4d5d3755d07db
Related-Bug: #1517010