Fetch latest ref from gerrit
In case there are a lot of commits to be collected together
it becomes hard to check before each build that all refs are
updated to the latest.

This commit adds function that will do it for you. It will
try to fetch for every value that is not looks like gerrit ref
(doesn't start with 'refs') the latest patchset number from
gerrit.

Change-Id: I4707dca474dbecb98783fa73765cdc862cbe7a58