Rebuild virtual envs for python jobs on master
VENV_CLEANUP bool parameter with value set to true by default is added
to the jobs configuration.
We don't need to take care of all venvs that tox creates as ones are
located in ${WORKSPACE} which is wiped out every job run.
So only removal of the venv directory created by virtualenv utility
outside of workspace is implemented.
Directory path for such environments is ${WORKSPACE}_VENV.

It is done in order to identify issues with pypi packages as soon
as possible. There's no such possibility in the meantime.

Change-Id: I4ebe08dbd90a1fe8995b344eaec0959c0a032c48
Related-Bug: #1477223