Add stable/proposed repo option for swarm tests
Tests runs included proposed repos for Ubuntu before. However it is
required to run tests only on stable repos on weekdays and on proposed
ones on weekends. For this purpose a new parameter ENABLE_PROPOSED was
added to swarm.runner and swarm.timer.* jobs. There are 2 timers: one
for weekdays (stable) and one for weekends (proposed).
run_system_test.sh was updated accordingly to form repo list depending
on ENABLE_PROPOSED property value.

Change-Id: Ia2c625848e8b06057f0684c3deb963b67aa99ffc
Closes-Bug: #1526338