mos 9.0 code sync: make it possible to specify a fallback branch
As of now stable/mitaka branches have not been cut in upstream yet.
It would be nice to be able to set a 'fallback' branch to be used
for syncing the code from while the given upstream branch does not
exist. I.e. we'll be syncing code from master until stable/mitaka
is cut in a particular project without having to reconfigure the
job for each project, when it is created.

Change-Id: Ia2785256d73048041c33ab7fb3fb9ecb249bac3e