Add separate recheck for Fuel CI
Gerrit Trigger plugin calculates result per event. So if we run only
partial recheck for unit-tests, deployment test results are ignored
and patch gets Verified+1 even if it fails deployment.

Thus we can only recheck all tests at once.

We setup a separate keyword for Fuel CI, so that we do not restart
heavy deployment tests when there are failures in lightweight upstream
gate.

This patch:

* adds recheck Fuel CI jobs including deployment tests by comment:

    fuel: recheck

* adds rechecks for previously skipped fuel-library unit tests.

Closes-Bug: #1545253

Change-Id: I3a65de39e26ace451ce1798d5087b7db9bf24124