mos: add jobs moving custom patches from one release to another
In MOS projects we have "custom" patches, that are not part of the
new OpenStack release we are consuming. When branches for a new MOS
release are created in downstream we have to re-propose these patches
to make sure we don't forget changes we need.

It works as follows. For each project:

  - find the common anscestor between $old and $new git branches
  - gather all the commits in ranges $ancestor..$old and
    $ancestor..$new
  - for each changeid present in $ancestor..$old, but not present
    in $ancestor..$new, do:

    - check out the corresponding git commit
    - upload it for review to $new branch skipping the rebase step,
      as it will most likely fail

Change-Id: Iaaec98271123a7c90416a8757a15509c1f0f156d