upstream code sync: run syncs more often till HCF
We don't really do active development in downstream, so all changes
(including bug fixes) go to upstream first and then are synced weekly
to downstream branches once merged.

At this point of release cycle we want to be more agile and not
wait for another week, till the recently merged fixes are synced by
the means of merge commits. Of course, jobs can be run on demand by
Infra team, but we don't need them to be constantly pinged by the
component teams, when people want to have a particular fix synced
right away.

Proposing merge commits daily sounds like a good compromise. If a
previous sync hasn't been merged by the time a new one is proposed,
the existing Gerrit review will be updated - no additional reviews
are created.

To be reverted on the HCF date.

Change-Id: I03ba07ccf5ddfd7b7715f79f49c6023527af0154