Fixed service_ha and service_ha_one_controller tests
Service-ha heat tests (autoscaling and waitcondition)
need more than 7gb ram for stable work.
In the test, we deployed three or more virtual machines
on that we need a small supply of RAM.
This commit fixed this issue. 
After recearch i see that older ram value= 7168mb in SLAVE_NODE_MEMORY
actually = 6977mb total RAM on real machine. 
A certain amount of memory is given to the OS. 
We use slaves with 64gb ram for this label 
and the jobs run exclusively on those slaves, so there shouldn't be any memory overcommit. 
I added +1024mb ram on slaves in this tests groups.

Change-Id: I6e5fb204e325952f3665e1a485e5af0ec87e3f42
Closes-Bug: #1563765