Move credentials to global defaults
Currently we use scm credentials hardcoded in macroses
That approach is not very clean because we often
copypaste these macroses w/o full understanding
of which credentials are use and why.
This approach is also unhandy for using with sandbox
because it requires seding lot of files.

This patch moves all such cases to global defaults and
adds verbosity in credentials usage.
We should continue to move all environment-specific
parameters to global defaults.
This will help to concentrate all properties which
should be changed in sandbox in one file.

Change-Id: Idd8834a3b24d582b23e2c3db6282b13b5e70c3ae