Drop massrebuild jobs
This patch removes massrebuild jobs because they
are otdated and haven't been launched for a long time.
pkgs-ci projects and layouts have changed significantly and
it's hard to determine if that these jobs work.
So now it's appropriate to drem them from repo and recreate
from scratch under related jira prod.

Change-Id: I6fff262fdd9eceb5ccde3817bbc24b4e0af83fef