Create baremetal acceptance jobs for 9.x
Fuel master will be set up on the VM using jenkins slave
with baremetal_acceptance label. On that slave we also store
devops templates for every particular environment (aplha and beta as for now)
All the following logic should be like for acceptance test, except for testrail
reporting, since we are uploading statistics for the part of the test suite only.

Related blueprint run-fuel-qa-on-baremetal

Change-Id: I9780c7220e49ba4a5f507aa11a7a3035d2ea7eb6