Custom BVT_2 test job without using Fuel ISO
This change includes template for custom BVT_2 job, builder script,
and project that realizes the job for 10.0.
Job requires URL to the fuel-release package and YAML files with
RPM and DEB repositories configuration, and accepts list of
fuel-qa gerrit changes to cherry-pick.

VENV_PATH is changed according to the new venv naming.

Change-Id: I0d01ca68d47826f9cd2bc833bddf8a9c2acea714