Use dedicated venv for upgrade tests
Upgrade tests are using several OS releases, which includes
upper constraints for python-clients and other packages.
Unfortunately we are not able to fix all compatibility issues
due do dep hell and drop of deprecates in upstream.
It's better to detach upgrade tests from swarm's venv.

We want to run 7-8 upgrades and 8-9 upgrades, and requirements
for 7.0 branch are incompatible with 8.0 branch's requirements.
The solution for this issue is reinstallation of requirements
each upgrade test run.
We can not run more that 1 env at the same time so
it shouldn't break anything.

Change-Id: I48b25637461a901b1590c0fd547ec2e7efd57ae4