Rework env_prefix generation for baremetal tests
We don't need to have unique env name for the baremetal tests.
Moreover we need to have it the same for one environment for the
tests to work, because we don't have many address pools for the public
network,for example, we are limited by our devops template.

Env name should be like baremetal.alpha for alpha, baremetal.beta for beta

Related blueprint run-fuel-qa-on-baremetal

Change-Id: I78f67c28d9cc76d4ba2a6edc1f917ae7b0cd33ae