Increase timeout for test jobs
We neeed to increase this timeout because otherwise those jobs are
aborted due to the fact that baremetal jobs run takes very long time
and we have only two environments. E.g. only two jobs could be run
at the time and all other jobs could wait in the queue for hours.

Related blueprint run-fuel-qa-on-baremetal

Change-Id: I0e84f674fd86526877b0928c64d0ccaa46ab2e8e