Add proposed for canary runs
There is a problem with systests on rpm packages.
We need to upgrade master node, but this is possible only
with packages published to proposed repo.
Proposed repo comes in list of extra repos for real pipelines but
it's omitted in case of canary run. This patch adds proposed
to list of default repos for enabling upgrade of master node with
latest fuel-qa code

Change-Id: I2ab6ae1570bc99fc843da3e00e6718d4861170e5
Cloases-Bug: #1618936