Add timers for 10.0 swarm
This patch adds timers for 10.0 swarm.
It's requested to run swarm on wednesdays.

So in this patch it's added timer for ubuntu stable on wed.
And just in case it will be also added timer for proposed on sun.

This patch reverts commit c4dfea8105375184fe39a4bb2365f0f62c34c53b
with some additional tuning

Change-Id: Ic5e92ff6b71abfce6dbd4da74866e6a83f7a20b2