[9.x][swarm] Direct usage of upstream CentOS
Since we are testing CentOS updates on a master node we have to install
and use our snapshots of upstream CentOS. Before the patch snapshots
have been installed via the mos-release package, which is inconvenient
since demands the new mos-release package each time we want to test
a new snapshot.

The patch installs snapshots of the os, extras and updates repositories
with the given snapshot id making use of the CENTOS_MIRROR_ID variable
which was always there but unused.

Change-Id: I1e331b6c292afbcf0bf063b02bedb34397a94bcb