[8.0][SWARM] Use snapshots of upstream CentOS mirrors
The commit introduces the ability to run SWARM tests with enabled
upstream CentOS repositories. For that reason a new variable
CENTOS_MIRROR_ID is added in SWARM tests. This variable is being used in
the run_system_test.sh builder to attach the "os", "updates" and "extras"
CentOS repositories to a Fuel master node on the first stage of each
system test. Also the snapshot job for 8.0 changed to provide the fresh
CentOS snapshot-id on each run.

Change-Id: I06d1f7470c98570773a8115b659b7dd5362531db